ros2_object_detection_components/pyproject.toml
[project]
name = "ros2_object_detection_components"
version = "0.0.0"
description = "A PyQT application used to collect datasets for training transporter networks."
authors = [
	{ name = "Peter David Fagan", email = "peterdavidfagan@gmail.com" }
]
readme = "README.md"

[tool.poetry]
package-mode=false
name = "ros2_object_detection_components"
version = "0.0.0"
description = "A PyQT application used to collect datasets for training transporter networks."
license = "Apache-2.0"
authors = [
	"Peter David Fagan <peterdavidfagan@gmail.com>"
]
readme = "README.md"
classifiers = [
    "Operating System :: POSIX :: Linux"
]

[tool.poetry.dependencies]
python = "3.10.6"
numpy = "^1.16.0"
scipy = "^1.13.0"
pillow = "10.3.0"
huggingface-hub = "^0.22.0"
torch = "2.3.0"
transformers = "4.41.1"
opencv-python = "4.6.0.66" 
opencv-contrib-python = "4.6.0.66" 
cvbridge3 = "^1.1"
catkin-pkg = "^1.0.0"
empy = "3.3.4"
lark = "^1.1.9"


[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
